changes are moved to new classes
new EntitySnapshotManagers are created and changes are moved to that new classes
