+++ test/Abp.Zero.SampleApp.Tests/EntityHistory/EntitySnapshotManager_Tests.cs
@@ -1,27 +1,28 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Abp.Domain.Repositories;
 using Abp.Domain.Uow;
-using Abp.EntityFramework;
 using Abp.EntityHistory;
-using Abp.ZeroCore.SampleApp.Core.EntityHistory;
+using Abp.Zero.SampleApp.EntityHistory;
 using Shouldly;
 using Xunit;
 
-namespace Abp.Zero.EntityHistory
+namespace Abp.Zero.SampleApp.Tests.EntityHistory
 {
-    public class EntityHistoryHelper_Tests : AbpZeroTestBase
+ public class EntitySnapshotManager_Tests : SampleAppTestBase
     {
         private readonly IRepository<UserTestEntity> _userRepository;
-        private IEntityHistoryHelper _entityHistoryHelper;
+        private IEntitySnapshotManager _entitySnapshotManager;
 
-        public EntityHistoryHelper_Tests()
+        public EntitySnapshotManager_Tests()
         {
             _userRepository = Resolve<IRepository<UserTestEntity>>();
-            _entityHistoryHelper = Resolve<IEntityHistoryHelper>();
+            _entitySnapshotManager = Resolve<IEntitySnapshotManager>();
+            Resolve<IEntityHistoryConfiguration>().IsEnabledForAnonymousUsers = true;
         }
 
         [Fact]
@@ -32,7 +33,7 @@ namespace Abp.Zero.EntityHistory
 
             using (var uow = Resolve<IUnitOfWorkManager>().Begin())
             {
-                var snapshot = await _entityHistoryHelper.GetEntitySnapshotAsync<UserTestEntity, int>(id, DateTime.Now);
+                var snapshot = await _entitySnapshotManager.GetSnapshotAsync<UserTestEntity, int>(id, DateTime.Now);
                 snapshot.ChangedPropertiesSnapshots.Count.ShouldBe(0);
                 snapshot.PropertyChangesStackTree.Count.ShouldBe(0);
             }
@@ -57,7 +58,7 @@ namespace Abp.Zero.EntityHistory
             using (var uow = Resolve<IUnitOfWorkManager>().Begin())
             {
                 //undo last update
-                var snapshot = await _entityHistoryHelper.GetEntitySnapshotAsync<UserTestEntity, int>(id, DateTime.Now.AddSeconds(-2));
+                var snapshot = await _entitySnapshotManager.GetSnapshotAsync<UserTestEntity, int>(id, DateTime.Now.AddSeconds(-2));
 
                 snapshot.ChangedPropertiesSnapshots.Count.ShouldBe(1);
                 snapshot.PropertyChangesStackTree.Count.ShouldBe(1);
@@ -68,7 +69,7 @@ namespace Abp.Zero.EntityHistory
 
 
                 //undo all changes
-                var snapshot2 = await _entityHistoryHelper.GetEntitySnapshotAsync<UserTestEntity>(id, DateTime.Now.AddDays(-1));
+                var snapshot2 = await _entitySnapshotManager.GetSnapshotAsync<UserTestEntity>(id, DateTime.Now.AddDays(-1));
 
                 snapshot2.ChangedPropertiesSnapshots.Count.ShouldBe(2);
                 snapshot2.PropertyChangesStackTree.Count.ShouldBe(2);
@@ -96,3 +97,5 @@ namespace Abp.Zero.EntityHistory
         }
     }
 }
+
+
